# studio/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ApiErrorIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
...
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CreateGenerationJobRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationBudgetRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "reason"
        ],
        "title": "CreateGenerationJobRequest",
        "type": "object"
      },
      "CreateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "collaboration_mode": {
            "default": "managed",
            "enum": [
              "managed",
              "co_design",
              "governance"
            ],
            "title": "Collaboration Mode",
            "type": "string"
          },
          "created_from": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Created From",
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "name",
          "actor",
          "reason"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "GenerationJobActionRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          }
        },
        "required": [
          "actor"
        ],
        "title": "GenerationJobActionRequest",
        "type": "object"
      },
...
      "ProductGenerationJobResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "budget": {
            "$ref": "#/components/schemas/GenerationBudget"
          },
          "checkpoint_step": {
            "anyOf": [
              {
                "enum": [
                  "prepare",
                  "generate",
                  "validate"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Step"
          },
          "consumed_bytes": {
            "title": "Consumed Bytes",
            "type": "integer"
          },
          "consumed_cost_units": {
            "title": "Consumed Cost Units",
            "type": "integer"
          },
          "consumed_duration_seconds": {
            "title": "Consumed Duration Seconds",
            "type": "number"
          },
          "consumed_files": {
            "title": "Consumed Files",
            "type": "integer"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "input_dependencies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Input Dependencies",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "const": "web_product",
            "title": "Kind",
            "type": "string"
          },
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationManifest"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/RevisionMetaResponse"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "provider": {
            "const": "deterministic_template",
            "title": "Provider",
            "type": "string"
          },
          "provider_version": {
            "title": "Provider Version",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "sandbox": {
            "$ref": "#/components/schemas/GenerationSandboxPolicy"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "paused",
              "succeeded",
              "failed",
              "stale_input",
              "budget_exhausted",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "web_generation_contract_revision_id": {
            "title": "Web Generation Contract Revision Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          },
          "workspace_relative_path": {
            "title": "Workspace Relative Path",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "revision_id",
          "meta",
          "project_id",
          "kind",
          "status",
          "provider",
          "provider_version",
          "input_dependencies",
          "web_generation_contract_revision_id",
          "workspace_id",
          "workspace_relative_path",
          "budget",
          "sandbox",
          "fingerprint",
          "attempt",
          "checkpoint_step",
          "consumed_files",
          "consumed_bytes",
          "consumed_duration_seconds",
          "consumed_cost_units",
          "manifest",
          "error_code",
          "error_summary",
          "content_hash"
        ],
        "title": "ProductGenerationJobResponse",
        "type": "object"
      },
...
      "ProductProjectResponse": {
        "additionalProperties": false,
        "properties": {
          "collaboration_mode": {
            "enum": [
              "managed",
              "co_design",
              "governance"
            ],
            "title": "Collaboration Mode",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_from": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Created From",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/RevisionMetaResponse"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "paused",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "revision_id",
          "meta",
          "name",
          "collaboration_mode",
          "status",
          "created_from",
          "content_hash"
        ],
        "title": "ProductProjectResponse",
        "type": "object"
      },
...
      "RevisionMetaResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "parent_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision Id"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "parent_revision_id",
          "created_at",
          "created_by",
          "reason"
        ],
        "title": "RevisionMetaResponse",
        "type": "object"
      },
      "SourceReference": {
        "additionalProperties": false,
        "description": "A compact pointer to the origin of a product-domain statement.",
        "properties": {
          "locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locator"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          },
          "source_id": {
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "user_input",
              "imported_artifact",
              "research",
              "evidence",
              "tool_result",
              "model_proposal",
              "human_decision"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "source_type",
          "source_id"
        ],
        "title": "SourceReference",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first command and projection API for Product Studio. Proposal and generation work is created as a persisted job; an explicit run command performs only the bounded local step. Generated source is not installed, executed, or exposed as a preview.",
    "title": "psyteardown Product Studio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Create Project",
        "tags": [
          "product-projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/generation-jobs": {
      "get": {
        "operationId": "list_generation_jobs_api_v1_projects__project_id__generation_jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductGenerationJobResponse"
                  },
                  "title": "Response List Generation Jobs Api V1 Projects  Project Id  Generation Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "List Generation Jobs",
        "tags": [
          "product-generation-jobs"
        ]
      },
      "post": {
        "operationId": "create_generation_job_api_v1_projects__project_id__generation_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGenerationJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Create Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/generation-jobs/{job_id}": {
      "get": {
        "operationId": "get_generation_job_api_v1_projects__project_id__generation_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Get Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/generation-jobs/{job_id}/cancellations": {
      "post": {
        "operationId": "cancel_generation_job_api_v1_projects__project_id__generation_jobs__job_id__cancellations_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Cancel Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/generation-jobs/{job_id}/pauses": {
      "post": {
        "operationId": "pause_generation_job_api_v1_projects__project_id__generation_jobs__job_id__pauses_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Pause Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/generation-jobs/{job_id}/resumes": {
      "post": {
        "operationId": "resume_generation_job_api_v1_projects__project_id__generation_jobs__job_id__resumes_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Resume Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/generation-jobs/{job_id}/retries": {
      "post": {
        "operationId": "retry_generation_job_api_v1_projects__project_id__generation_jobs__job_id__retries_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Retry Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/generation-jobs/{job_id}/runs": {
      "post": {
        "operationId": "run_generation_job_api_v1_projects__project_id__generation_jobs__job_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductGenerationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Run Generation Job",
        "tags": [
          "product-generation-jobs"
        ]
      }
    },
...
  }
}